Refactor citation modules and add documentation to citation_formatter.rb
